requirements.txt
langchain-community==0.0.38
chromadb==0.5.0
sentence-transformers==3.0.1
pypdf2==3.0.1
python-dotenv==1.0.0
spacy==3.7.5
streamlit==1.38.0
fastapi==0.115.0
